added if staatements to migration

+++ db/migrate/20250325222319_remove_encrypted_password_from_business.rb
@@ -1,9 +1,19 @@
 class RemoveEncryptedPasswordFromBusiness < ActiveRecord::Migration[7.2]
   def change
-    remove_column :businesses, :encrypted_password
-    remove_column :businesses, :reset_password_token
-    remove_column :businesses, :remember_created_at
-    remove_column :businesses, :reset_password_sent_at
-    remove_column :businesses, :email
+    if column_exists? :businesses, :encrypted_password
+      remove_column :businesses, :encrypted_password
+    end
+    if column_exists? :businesses, :reset_password_token
+      remove_column :businesses, :reset_password_token
+    end
+    if column_exists? :businesses, :remember_created_at
+      remove_column :businesses, :remember_created_at
+    end
+    if column_exists? :businesses, :reset_password_sent_at
+      remove_column :businesses, :reset_password_sent_at
+    end
+    if column_exists? :businesses, :email
+      remove_column :businesses, :email
+    end
   end
 end
